xxd: reliably name the variable independent of build root path
As xxd uses the path name to name the variable, and the source hardcodes
the assumption that the build dir is exactly one level below the project
root, having that at a different place breaks the build. To circumvent
this issue, the variable name is passed as an extra arg to the
generator.